refactor: portal animated combobox
